init VideoWindowDisplay after registering defaults
It was being created before the applicationDidFinishLaunching delegate
was called (which was setting up the default preferences). This meant
that when there was no UserDefault saved, the VideoWindowsDisplay would
look for the hard-coded default (which hadn't been set yet) and thus
got 0 for all preferences (notably making the video windows
transparent). Interestingly, if the user sets a preference to be the
same as the registered default, it actually gets removed from the
physical file since its the same as the hard-coded default preference.

Closes #10